requirements.txt
tensorflow
tensorflow-hub
librosa
numpy
soundfile
twilio
python-dotenv
flask
requests